+++ src/entities/study/overview/api/benefits.api.ts
@@ -1,19 +1,11 @@
-export type {
-  EditRuleRequest,
-  EditRuleResponse,
-  EditStudyRule,
-  GetStudyRulesResponse,
-  StudyRule,
-} from "@/entities/study/overview/api/rules.api.type";
-
 import { createJsonRequestInit, fetchAPI } from "@/shared/api/util/fetcher";
 
 import { API_ENDPOINT } from "@/shared/api/lib";
 
 import {
+  EditBenefitRequest,
+  EditBenefitResponse,
   GetStudyBenefitsResponse,
-  PostBenefitRequest,
-  PostBenefitResponse,
 } from "@/entities/study/overview/api/benefits.api.type";
 
 export async function getBenefits(studyId: string) {
@@ -23,10 +15,10 @@ export async function getBenefits(studyId: string) {
   return response;
 }
 
-export async function postBenefit({ studyId, benefits }: PostBenefitRequest) {
+export async function editBenefits({ studyId, benefits }: EditBenefitRequest) {
   const { url, method } = API_ENDPOINT.benefits.postStudyBenefits(studyId);
 
-  return await fetchAPI<PostBenefitResponse[]>(
+  return await fetchAPI<EditBenefitResponse[]>(
     url,
     createJsonRequestInit(method, benefits)
   );
