+++ src/entities/study/overview/api/rules.api.ts
@@ -1,31 +1,22 @@
-export type {
-  EditRuleRequest,
-  EditRuleResponse,
-  EditStudyRule,
-  GetStudyRulesResponse,
-  StudyRule,
-} from "@/entities/study/overview/api/rules.api.type";
-
 import { createJsonRequestInit, fetchAPI } from "@/shared/api/util/fetcher";
 
 import {
-  EditRuleRequest,
-  EditRuleResponse,
+  EditRulesResponse,
   GetStudyRulesResponse,
-  StudyRule,
 } from "@/entities/study/overview/api/rules.api.type";
+import { EditRulesRequest } from "@/features/study-room/types/study-room.api";
 import { API_ENDPOINT } from "@/shared/api/lib";
 
-export async function getRules(studyId: StudyRule["id"]) {
+export async function getRules(studyId: string) {
   const response = await fetchAPI<GetStudyRulesResponse>(
     API_ENDPOINT.rules.getStudyRules(studyId).url
   );
   return response;
 }
 
-export async function editRules({ studyId, rules }: EditRuleRequest) {
+export async function editRules({ studyId, rules }: EditRulesRequest) {
   const { url, method } = API_ENDPOINT.rules.postStudyRules(studyId);
-  return await fetchAPI<EditRuleResponse>(
+  return await fetchAPI<EditRulesResponse>(
     url,
     createJsonRequestInit(method, rules)
   );
