+++ src/entities/study/overview/api/rules.api.type.ts
@@ -1,17 +1,17 @@
+import { RuleInterface } from "@/features/study-room/types/study-room.type";
+
 // Study 규칙
-export interface StudyRule {
+export interface Rule {
   id: number;
   ruleNum: number;
   content: string;
 }
 
-export type GetStudyRulesResponse = StudyRule[];
-
-export type EditStudyRule = Omit<StudyRule, "id">;
+export type GetStudyRulesResponse = Rule[];
 
 export type EditRuleRequest = {
-  studyId: string; // id로 바꾸면 되긴함
-  rules: EditStudyRule;
+  studyId: string;
+  rules: RuleInterface[];
 };
 
-export type EditRuleResponse = EditStudyRule[];
+export type EditRulesResponse = Rule[];
