+++ src/features/recruitment-status/ui/recruitment-status.ui.type.ts
