+++ src/features/share/ui/share.ui.type.ts
