+++ src/shared/api/lib/rules.ts
@@ -1,7 +1,7 @@
 import { Method } from "@/shared/api/util/fetcher";
 
 export const rules = {
-  getStudyRules: (studyId: number) => ({
+  getStudyRules: (studyId: string) => ({
     url: `${process.env.NEXT_PUBLIC_BASE_URL}/studies/${studyId}/rules`,
     method: Method.GET,
   }),
